pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monday-exporter"
version = "0.1.0"
description = "CLI tool to export Monday.com board data into formatted Excel workbooks."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Your Name" }
]
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "requests>=2.32.0",
  "openpyxl>=3.1.5",
  "pydantic>=2.9.2",
  "python-dateutil>=2.9.0.post0",
  "rich>=13.9.1",
  "typer>=0.12.5"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.5",
  "mypy>=1.12.0",
  "types-requests>=2.32.0",
  "types-python-dateutil>=2.9.0"
]

[project.scripts]
monday-export = "monday_exporter.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
